Contractを削除
